+++ packages/react-notifications/lib/react-notifications.js
@@ -16,32 +16,44 @@ export class Notifications extends Component {
 		this.setState({
 			hide: true
 		});
+		this.props.cbFunction ? this.props.cbFunction() : '';
 	}
 
 	render() {
 		const { hide } = this.state;
-		
+
 		let classMain = "sui-notice";
 		let classIcon = "sui-notice-icon sui-md";
 
-		switch (this.props.type) {
+		switch ( this.props.type ) {
 			case "info":
 			case "success":
 			case "warning":
 			case "error":
 			case "upsell":
 				classMain += " sui-notice-" + this.props.type;
-				classIcon += " sui-icon-info";
-				break;
-
-			case "loading":
-				classIcon += " sui-icon-loader sui-loading";
+				if ( this.props.loading ) {
+					classIcon += " sui-icon-loader sui-loading";
+				} else {
+					classIcon += " sui-icon-info";
+				}
 				break;
 
 			default:
-				classIcon += " sui-icon-info";
+				if ( this.props.loading ) {
+					classIcon += " sui-icon-loader sui-loading";
+				} else {
+					classIcon += " sui-icon-info";
+				}
 				break;
 		}
+
+		const lang = Object.assign(
+			{
+				dismiss: 'Hide Notification'
+			},
+			this.props.sourceLang
+		);
 
 		const message = (
 			<div className="sui-notice-message">
@@ -54,7 +66,7 @@ export class Notifications extends Component {
 			<div className="sui-notice-actions">
 				<ButtonIcon
 					icon="check"
-					label="Hide"
+					label={ lang.dismiss }
 					onClick={ e => this.close(e) }
 				/>
 			</div>