+++ CHANGELOG.md
@@ -2,6 +2,84 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org/) for commit guidelines.
 
+## [1.8.1](https://github.com/wpmudev/shared-ui/compare/v1.8.0...v1.8.1) (2022-05-03)
+
+#### ✨ New Features
+* `react-input`
+  * Add `prefix` prop. [#207](https://github.com/wpmudev/shared-ui-react/pull/207) ([SUI-356](https://incsub.atlassian.net/browse/SUI-356)) ([@creador-dev](https://github.com/creador-dev))
+  * Add `suffix` prop. [#206](https://github.com/wpmudev/shared-ui-react/pull/206) ([SUI-357](https://incsub.atlassian.net/browse/SUI-357)) ([@creador-dev](https://github.com/creador-dev))
+  * Allow constrained field. [#205](https://github.com/wpmudev/shared-ui-react/pull/205) ([SUI-358](https://incsub.atlassian.net/browse/SUI-358)) ([@iamleigh](https://github.com/iamleigh))
+* `react-notifications`
+  * Support callback function for dismiss button. [#187](https://github.com/wpmudev/shared-ui-react/pull/187) ([SUI-349](https://incsub.atlassian.net/browse/SUI-349)) ([@creador-dev](https://github.com/creador-dev))
+
+#### 🚀 Improvements
+* `react-dropdown`
+  * Add `disabled` property for trigger menu button. [#184](https://github.com/wpmudev/shared-ui-react/pull/184) ([SUI-341](https://incsub.atlassian.net/browse/SUI-341)) ([@creador-dev](https://github.com/creador-dev))
+* `react-notifications`
+  * Ability to edit dismiss button label. [#181](https://github.com/wpmudev/shared-ui-react/pull/181) ([SUI-350](https://incsub.atlassian.net/browse/SUI-350)) ([@iamleigh](https://github.com/iamleigh))
+  * Loading must be a state of the component. [#180](https://github.com/wpmudev/shared-ui-react/pull/180) ([SUI-347](https://incsub.atlassian.net/browse/SUI-347)) ([@iamleigh](https://github.com/iamleigh))
+
+#### 🐛 Bug Fixes
+* `react-pagination`
+  * Parent wrapper breaks the component. [#209](https://github.com/wpmudev/shared-ui-react/pull/209) ([SUI-362](https://incsub.atlassian.net/browse/SUI-362)) ([@creador-dev](https://github.com/creador-dev))
+  * Wrong package version for `builder`. [#198](https://github.com/wpmudev/shared-ui-react/pull/198) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `react-post`
+  * Image keeps on loading when is empty or null. [#183](https://github.com/wpmudev/shared-ui-react/pull/183) ([SUI-343](https://incsub.atlassian.net/browse/SUI-343)) ([@creador-dev](https://github.com/creador-dev))
+  * Wrong package version for `builder`. [#199](https://github.com/wpmudev/shared-ui-react/pull/199) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+  * Disabled state is missing. [#185](https://github.com/wpmudev/shared-ui-react/pull/185) ([SUI-345](https://incsub.atlassian.net/browse/SUI-345)) ([@creador-dev](https://github.com/creador-dev))
+  * Post title HTML chars are rendered encoded on hovering. [#163](https://github.com/wpmudev/shared-ui-react/pull/163) ([SUI-264](https://incsub.atlassian.net/browse/SUI-264)) ([@creador-dev](https://github.com/creador-dev))
+* `react-progress-bar`
+  * Wrong package version for `builder`. [#200](https://github.com/wpmudev/shared-ui-react/pull/200) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `react-notifications`
+  * Wrong package version for `builder`. [#197](https://github.com/wpmudev/shared-ui-react/pull/197) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `react-modal`
+  * Wrong package version for `builder`. [#196](https://github.com/wpmudev/shared-ui-react/pull/196) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `react-input`
+  * Wrong package version for `builder`. [#195](https://github.com/wpmudev/shared-ui-react/pull/195) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `react-dropdown`
+  * Wrong package version for `builder`. [#194](https://github.com/wpmudev/shared-ui-react/pull/194) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+  * Missing `disabled` styles for options. [#186](https://github.com/wpmudev/shared-ui-react/pull/186) ([SUI-342](https://incsub.atlassian.net/browse/SUI-342)) ([@creador-dev](https://github.com/creador-dev))
+* `react-button`
+  * Wrong package version for `builder`. [#193](https://github.com/wpmudev/shared-ui-react/pull/193) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `react-box`
+  * Wrong package version for `builder`. [#192](https://github.com/wpmudev/shared-ui-react/pull/192) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+* `core`
+  * Wrong package version for `builder`. [#190](https://github.com/wpmudev/shared-ui-react/pull/190) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+  * Missing `ProgressBar` component. [#182](https://github.com/wpmudev/shared-ui-react/pull/182) ([SUI-351](https://incsub.atlassian.net/browse/SUI-351)) ([@iamleigh](https://github.com/iamleigh))
+* `react-accordion`
+  * Wrong package version for `builder`. [#191](https://github.com/wpmudev/shared-ui-react/pull/191) ([SUI-353](https://incsub.atlassian.net/browse/SUI-353)) ([@creador-dev](https://github.com/creador-dev))
+  * Demo image is missing in the stories. [#174](https://github.com/wpmudev/shared-ui-react/pull/174) ([SUI-335](https://incsub.atlassian.net/browse/SUI-335)) ([@creador-dev](https://github.com/creador-dev))
+
+#### 📝 Documentation
+* `react-pagination`
+  * Improve stories. [#213](https://github.com/wpmudev/shared-ui-react/pull/213) ([SUI-339](https://incsub.atlassian.net/browse/SUI-339)) ([@iamleigh](https://github.com/iamleigh))
+* `react-accordion`
+  * Improve stories. [#208](https://github.com/wpmudev/shared-ui-react/pull/208) ([SUI-352](https://incsub.atlassian.net/browse/SUI-352)) ([@iamleigh](https://github.com/iamleigh))
+* `react-progress-bar`
+  * Improve stories. [#176](https://github.com/wpmudev/shared-ui-react/pull/176) ([SUI-337](https://incsub.atlassian.net/browse/SUI-337)) ([@creador-dev](https://github.com/creador-dev))
+* `react-input`
+  * Improve stories. [#175](https://github.com/wpmudev/shared-ui-react/pull/175) ([SUI-336](https://incsub.atlassian.net/browse/SUI-336)) ([@iamleigh](https://github.com/iamleigh))
+* `react-notifications`
+  * Improve stories. [#179](https://github.com/wpmudev/shared-ui-react/pull/179) ([SUI-346](https://incsub.atlassian.net/browse/SUI-346)) ([@iamleigh](https://github.com/iamleigh))
+* `react-post`
+  * Improve stories. [#178](https://github.com/wpmudev/shared-ui-react/pull/178) ([SUI-340](https://incsub.atlassian.net/browse/SUI-340)) ([@iamleigh](https://github.com/iamleigh))
+* `react-dropdown`
+  * Improve stories. [#177](https://github.com/wpmudev/shared-ui-react/pull/177) ([SUI-338](https://incsub.atlassian.net/browse/SUI-338)) ([@iamleigh](https://github.com/iamleigh))
+* `react-button-icon`
+  * Improve stories. [#173](https://github.com/wpmudev/shared-ui-react/pull/173) ([SUI-334](https://incsub.atlassian.net/browse/SUI-334)) ([@iamleigh](https://github.com/iamleigh))
+* `react-button`
+  * Improve stories. [#172](https://github.com/wpmudev/shared-ui-react/pull/172) ([SUI-333](https://incsub.atlassian.net/browse/SUI-333)) ([@iamleigh](https://github.com/iamleigh))
+
+#### 🏠 Internal
+* Invalid hook call. [#204](https://github.com/wpmudev/shared-ui-react/pull/204) ([SUI-355](https://incsub.atlassian.net/browse/SUI-355)) ([@creador-dev](https://github.com/creador-dev))
+* Set `react` dependency to latest version for all packages. [#171](https://github.com/wpmudev/shared-ui-react/pull/171) ([SUI-203](https://incsub.atlassian.net/browse/SUI-203)) ([@iamleigh](https://github.com/iamleigh))
+* Rollup was trying to import and compile `react-dom` causing an error during build. [#166](https://github.com/wpmudev/shared-ui-react/pull/166) ([SUI-329](https://incsub.atlassian.net/browse/SUI-329)) ([@creador-dev](https://github.com/creador-dev))
+* Upgrade `SUI` dependency to its latest version. [#160](https://github.com/wpmudev/shared-ui-react/pull/160) ([SUI-314](https://incsub.atlassian.net/browse/SUI-314)) ([@iamleigh](https://github.com/iamleigh))
+
+#### Committers: 2
+- Leighton Sapir ([@iamleigh](https://github.com/iamleigh))
+- Pawan Kumar ([@creador-dev](https://github.com/creador-dev))
+
 ## [1.8.0](https://github.com/wpmudev/shared-ui/compare/v1.7.0...v1.8.0) (2022-01-30)
 
 #### ✨ New Features